==> 1_preprocessing.ipynb <==
# Packages

import pandas as pd

## 1. Publications

There are a total of 1959 publications.

df_wos = pd.read_csv('data/publications.csv', sep=',', encoding='UTF-8', dtype={'Publication Date':'Int32'})
df_wos.shape

91% are papers and 7% reviews.

df_wos[['Document Type']].groupby('Document Type').size()

100*df_wos[['Document Type']].groupby('Document Type').size()/df_wos.shape[0]

Most publications are from the past 5 years.

df_wos[['Publication Date']].groupby('Publication Date').size().plot.line(style='.-')

Almost all publications (n=1928) include DOI.

sum(-df_wos.DOI.isna())

## 2. Altmetric outputs

1213 publications (63% with DOI) are indexed in Altmetric.com and 1097 have at least one altmetric mention.

df_ao = pd.read_csv('data/altmetric_outputs.csv', sep=',', encoding='UTF-8')
df_ao.shape

df_ao[df_ao['Altmetric Attention Score']>0].shape

Most of the publications with altmetric attention (n=1047) have been mentioned in one of the four sources studied. 

df_ao[(df_ao['Twitter mentions']>0) | (df_ao['News mentions']>0) | (df_ao['Policy mentions']>0) | (df_ao['Wikipedia mentions']>0)].shape

## 3. Altmetric mentions

These publications have 14,919 altmetric mentions.

df_am = pd.read_csv('data/altmetric_mentions.csv', sep=',',  encoding='UTF-8', dtype={'External Mention ID':'string'})
df_am.shape

83% are Twitter mentions, 10% news mentions, 1% policy mentions, and 1% Wikipedia mentions.

df_am[['Mention Type']].groupby('Mention Type').size()

100*df_am[['Mention Type']].groupby('Mention Type').size()/df_am.shape[0]

df_am[df_am['Mention Type']=='Tweet'].to_csv('data/twitter_mentions.tsv', sep='\t', index=False, encoding='UTF-8')

df_am[df_am['Mention Type']=='News story'].to_csv('data/news_mentions.tsv', sep='\t', index=False, encoding='UTF-8')

df_am[df_am['Mention Type']=='Wikipedia page'].to_csv('data/wikipedia_mentions.tsv', sep='\t', index=False, encoding='UTF-8')
