==> 2_twitter_metadata.ipynb <==
# Packages

import pandas as pd
import botometer
import tweepy
from functions_alt import tweeters_metadata, tweets_metadata, twitter_bots

# 1. Altmetric mentions

There are 12,413 Twitter mentions.

df = pd.read_csv('data/twitter_mentions.tsv', sep='\t', encoding='UTF-8', dtype={'Outlet or Author':'string', 'External Mention ID':'string'})
df.shape

These tweets are published by 7391 unique tweeters.

df[['Outlet or Author']].drop_duplicates().shape

# 2. Twitter API metadata

auth = tweepy.AppAuthHandler(consumer_key='',
                             consumer_secret='')
client = tweepy.API(auth, wait_on_rate_limit=True,
                   parser=tweepy.parsers.JSONParser())

## 2.1. Tweeters

fields_to_retrieve = ['id', 'id_str', 'name', 'screen_name', 'location', 'description',
                      'url', 'protected', 'followers_count', 'friends_count',
                      'listed_count', 'created_at', 'favourites_count',
                      'verified', 'statuses_count', 'lang', 'is_translator',
                      'is_translation_enabled', 'default_profile', 'default_profile_image']

tweeters = tweeters_metadata(fields_to_retrieve, df[['Outlet or Author']].drop_duplicates(), client)

tweeters.shape

tweeters.to_csv('data/Twitter/tweeters_metadata.tsv', sep='\t', index=False, encoding='UTF-8', escapechar='\\')

## 2.2. Tweets

fields_to_retrieve = ['created_at', 'id', 'id_str', 'text', 'source',
                      'in_reply_to_status_id_str', 'in_reply_to_user_id_str', 'is_quote_status', 'retweet_count', 'favorite_count', 'retweeted_status.id', 'retweeted',
                      'lang', 'geo', 'place', 'contributors', 
                      'user.id', 'user.id_str']

tweets = tweets_metadata(fields_to_retrieve, df[['External Mention ID']].drop_duplicates(), client)

tweets.shape

tweets.to_csv('data/Twitter/tweets_metadata.tsv', sep='\t', index=False, encoding='UTF-8', escapechar='\\')

# 3. Botometer API

The Botometer API query was chunked into 4 parts of 2000 accounts each.

rapidapi_key = ''
twitter_app_auth = {
    'consumer_key': '',
    'consumer_secret': '',
    'access_token': '',
    'access_token_secret': '',
  }
  
blt_twitter = botometer.Botometer(wait_on_ratelimit=True, rapidapi_key=rapidapi_key, **twitter_app_auth)

tweeters = pd.read_csv('data/Twitter/tweeters_metadata.tsv', sep='\t', encoding='UTF-8', escapechar='\\', dtype={'id_str':'string'})

#tw_bots_1 = twitter_bots(tweeters['id_str'].tolist()[0:2000], blt_twitter)
#tw_bots_2 = twitter_bots(tweeters['id_str'].tolist()[2000:4000], blt_twitter)
#tw_bots_3 = twitter_bots(tweeters['id_str'].tolist()[4000:6000], blt_twitter)
tw_bots_4 = twitter_bots(tweeters['id_str'].tolist()[6000:7309], blt_twitter)

#tw_bots_1.to_csv('data/Twitter/tweeters_bots_1.tsv', sep='\t', index=False, encoding='UTF-8')
#tw_bots_2.to_csv('data/Twitter/tweeters_bots_2.tsv', sep='\t', index=False, encoding='UTF-8')
#tw_bots_3.to_csv('data/Twitter/tweeters_bots_3.tsv', sep='\t', index=False, encoding='UTF-8')
tw_bots_4.to_csv('data/Twitter/tweeters_bots_4.tsv', sep='\t', index=False, encoding='UTF-8')
