==> 3_news.ipynb <==
# Packages

import pandas as pd
from functions_alt import news_data, full_url

# 1. Altmetric mentions

df_am = pd.read_csv('data/altmetric_mentions.csv', sep=',', encoding='UTF-8')

There are 1513 news mentions.

df_am = df_am[df_am['Mention Type']=='News story'].reset_index(drop=True)
df_am.shape

But, 139 do not include the news media URL, so they are removed.

df_am = df_am[~df_am['Mention URL'].isna()].reset_index(drop=True)
df_am.shape

# 2. Site Traffic API

df_am['Clean URL'] = df_am['Mention URL']

df_am['Clean URL'] = df_am['Clean URL'].str.replace('(http://)|(https://)|(www.)', '', regex=True)
df_am['Clean URL'] = df_am['Clean URL'].str.replace('/.*', '', regex=True)

There are 447 news media.

len(list(set(df_am['Clean URL'].tolist())))

## 2.1. API query 1

All URLs.

url = 'https://zylalabs.com/api/29/site+traffic+api/93/traffic+source+and+overview'
key = ''
header = {'Authorization':key}

df_news = news_data(url, header, list(set(df_am['Clean URL'].tolist())))

df_news.to_csv('data/News/media_metrics.tsv', sep='\t', encoding='UTF-8', index=False)

## 2.2. API query 2

Short URLs are fixed.

df_am['Full URL'] = None
df_am.loc[df_am['Clean URL']=='ct.moreover.com', 'Full URL'] = df_am.loc[df_am['Clean URL']=='ct.moreover.com', 'Mention URL']

for i in df_am[df_am['Clean URL']=='ct.moreover.com'].index:
    df_am.loc[i, 'Full URL'] = full_url(df_am.loc[i, 'Mention URL'])

df_am['Full URL'] = df_am['Full URL'].str.replace('(http://)|(https://)|(www.)', '', regex=True)
df_am['Full URL'] = df_am['Full URL'].str.replace('/.*', '', regex=True)

df_am.loc[df_am['Clean URL']=='ct.moreover.com', 'Full URL']

len(list(set(df_am['Full URL'].tolist())))

len(set(df_am[df_am['Full URL'].isin(df_am['Clean URL'])]['Full URL']))

len(set(df_am[(~df_am['Full URL'].isin(df_am['Clean URL'])) & (~df_am['Full URL'].isna())]['Full URL']))

url = 'https://zylalabs.com/api/29/site+traffic+api/93/traffic+source+and+overview'
key = ''
header = {'Authorization':key}

df_news_full = news_data(url, header, list(set(df_am[(~df_am['Full URL'].isin(df_am['Clean URL'])) & (~df_am['Full URL'].isna())]['Full URL'])))

df_news_full.to_csv('data/News/media_metrics_full.tsv', sep='\t', encoding='UTF-8', index=False)

df_am.loc[df_am['Clean URL']=='ct.moreover.com', 'Clean URL'] = df_am.loc[df_am['Clean URL']=='ct.moreover.com', 'Full URL']

df_am.to_csv('data/News/news_mentions_full.tsv', sep='\t', index=False, encoding='UTF-8')
