==> 6_news_analysis.ipynb <==
# Packages

import pandas as pd

# 1. Data import

## 1.1. Mentions

df = pd.read_csv('data/News/news_mentions_full.tsv', sep='\t', encoding='UTF-8', dtype={'pageid':'str'})
df.shape

df.rename({'Clean URL':'SiteName'}, axis=1, inplace=True)

df = df[df['SiteName']!='ct.moreover.com'].copy()
df.shape

## 1.2. News media

df_pages_1 = pd.read_csv('data/News/media_metrics.tsv', sep='\t', encoding='UTF-8', dtype={'Wikipedia_page_id':'str'})
df_pages_1 = df_pages_1[(~df_pages_1['Category'].isna())].reset_index(drop=True)
df_pages_1.shape

df_pages_2 = pd.read_csv('data/News/media_metrics_full.tsv', sep='\t', encoding='UTF-8', dtype={'Wikipedia_page_id':'str'})
df_pages_2 = df_pages_2[(~df_pages_2['Category'].isna())].reset_index(drop=True)
df_pages_2.shape

df_pages_1[df_pages_1['SiteName'].isin(df_pages_2.SiteName)].shape

df_pages = pd.concat([df_pages_1, df_pages_2])
df_pages = df_pages.drop_duplicates().reset_index(drop=True)

49 mentions/27 websites have not metadata

df[~df['SiteName'].isin(df_pages['SiteName'].tolist())].shape

len(set(df[~df['SiteName'].isin(df_pages['SiteName'].tolist())]['SiteName']))

df[df['SiteName'].isin(df_pages.SiteName)].shape

df_pages['EstimatedMonthlyVisits'] = df_pages['EstimatedMonthlyVisits.2022-11-01']+df_pages['EstimatedMonthlyVisits.2022-12-01']+df_pages['EstimatedMonthlyVisits.2023-01-01']
df_pages['EstimatedMonthlyVisits'] = df_pages['EstimatedMonthlyVisits']/1000000

df_pages['main_category'] = df_pages['Category']

for x in range(df_pages.shape[0]):
    try:
        df_pages.loc[x, 'main_category'] = df_pages.loc[x, 'Category'].split('/')[0]
    except:
        pass

df_pages['main_category'].value_counts()

Minor categories are aggregated.

df_pages.loc[~df_pages['main_category'].isin(['news_and_media', 'arts_and_entertainment',
                                              'science_and_education', 'health',
                                              'computers_electronics_and_technology', 'finance']),
            'main_category'] = 'Other'

df_pages['main_category'].value_counts()

df_pages['Engagments.Visits'] = df_pages['Engagments.Visits']/1000000

df_pages['Local'] = 0
df_pages.loc[df_pages['topcountry_name']==724, 'Local'] = 1

# 2. Analysis

df_pages.groupby('main_category').agg(Websites=('main_category', 'count'),
                                      Local=('Local', 'sum'),
                                      Three_Visits=('EstimatedMonthlyVisits', 'mean'),
                                      Visits=('Engagments.Visits', 'mean'),
                                      Time_Visits=('Engagments.TimeOnSite', 'mean'),
                                      Page_Visit=('Engagments.PagePerVisit', 'mean'),
                                      Small=('IsSmall', 'sum'),
                                      Rank=('GlobalRank.Rank', 'mean'),
                                      Bounce=('Engagments.BounceRate', 'mean'),
                                      Search=('TrafficSources.Search', 'mean'),
                                      Direct=('TrafficSources.Direct', 'mean'),
                                      Social=('TrafficSources.Social', 'mean')).round(2).reset_index().sort_values(by='Websites', ascending=False)

df_pages.groupby(lambda x: True).agg(Websites=('main_category', 'count'),
                                     Local=('Local', 'sum'),
                                     Three_Visits=('EstimatedMonthlyVisits', 'mean'),
                                     Visits=('Engagments.Visits', 'mean'),
                                     Time_Visits=('Engagments.TimeOnSite', 'mean'),
                                     Page_Visit=('Engagments.PagePerVisit', 'mean'),
                                     Small=('IsSmall', 'sum'),
                                     Rank=('GlobalRank.Rank', 'mean'),
                                     Bounce=('Engagments.BounceRate', 'mean'),
                                     Search=('TrafficSources.Search', 'mean'),
                                     Direct=('TrafficSources.Direct', 'mean'),
                                     Social=('TrafficSources.Social', 'mean')).round(2).reset_index().sort_values(by='Websites', ascending=False)

df_pages.groupby('main_category').agg(Websites=('main_category', 'count'),
                                      Local=('Local', 'sum'),
                                      Three_Visits=('EstimatedMonthlyVisits', 'median'),
                                      Visits=('Engagments.Visits', 'median'),
                                      Time_Visits=('Engagments.TimeOnSite', 'median'),
                                      Page_Visit=('Engagments.PagePerVisit', 'median'),
                                      Small=('IsSmall', 'sum'),
                                      Rank=('GlobalRank.Rank', 'median'),
                                      Bounce=('Engagments.BounceRate', 'median'),
                                      Search=('TrafficSources.Search', 'median'),
                                      Direct=('TrafficSources.Direct', 'median'),
                                      Social=('TrafficSources.Social', 'median')).round(2).reset_index().sort_values(by='Websites', ascending=False)

df_pages.groupby(lambda x: True).agg(Websites=('main_category', 'count'),
                                     Local=('Local', 'sum'),
                                     Three_Visits=('EstimatedMonthlyVisits', 'median'),
                                     Visits=('Engagments.Visits', 'median'),
                                     Time_Visits=('Engagments.TimeOnSite', 'median'),
                                     Page_Visit=('Engagments.PagePerVisit', 'median'),
                                     Small=('IsSmall', 'sum'),
                                     Rank=('GlobalRank.Rank', 'median'),
                                     Bounce=('Engagments.BounceRate', 'median'),
                                     Search=('TrafficSources.Search', 'median'),
                                     Direct=('TrafficSources.Direct', 'median'),
                                     Social=('TrafficSources.Social', 'median')).round(2).reset_index().sort_values(by='Websites', ascending=False)

df_pages_m = df.merge(df_pages, how='inner', on='SiteName')

df_pages_m.groupby('main_category').agg(Mentions=('main_category', 'count'),
                                        Local=('Local', 'sum'),
                                        Papers=('Details Page URL', 'nunique'),
                                        Three_Visits=('EstimatedMonthlyVisits', 'mean'),
                                        Visits=('Engagments.Visits', 'mean'),
                                        Time_Visits=('Engagments.TimeOnSite', 'mean'),
                                        Page_Visit=('Engagments.PagePerVisit', 'mean'),
                                        Small=('IsSmall', 'sum'),
                                        Rank=('GlobalRank.Rank', 'mean'),
                                        Bounce=('Engagments.BounceRate', 'mean'),
                                        Search=('TrafficSources.Search', 'mean'),
                                        Direct=('TrafficSources.Direct', 'mean'),
                                        Social=('TrafficSources.Social', 'mean')).round(2).reset_index()

df_pages_m.groupby(lambda x: True).agg(Mentions=('main_category', 'count'),
                                       Local=('Local', 'sum'),
                                       Papers=('Details Page URL', 'nunique'),
                                       Three_Visits=('EstimatedMonthlyVisits', 'mean'),
                                       Visits=('Engagments.Visits', 'mean'),
                                       Time_Visits=('Engagments.TimeOnSite', 'mean'),
                                       Page_Visit=('Engagments.PagePerVisit', 'mean'),
                                       Small=('IsSmall', 'sum'),
                                       Rank=('GlobalRank.Rank', 'mean'),
                                       Bounce=('Engagments.BounceRate', 'mean'),
                                       Search=('TrafficSources.Search', 'mean'),
                                       Direct=('TrafficSources.Direct', 'mean'),
                                       Social=('TrafficSources.Social', 'mean')).round(2).reset_index()

df_mentions = df_pages_m.groupby(['Details Page URL', 'main_category']).agg(Mentions=('main_category', 'count'),
                                                                            Avg_visits=('Engagments.Visits', 'mean'),
                                                                            Avg_time=('Engagments.TimeOnSite', 'mean'),
                                                                            Avg_pages=('Engagments.PagePerVisit', 'mean'),
                                                                            Avg_rank=('GlobalRank.Rank', 'mean')).round(2).reset_index()

df_mentions.to_csv('data/News/publications_metrics_mean.tsv', sep='\t', index=False)

df_pages.loc[df_pages['main_category']!='news_and_media', 'main_category'] = 'specialized_media'

df_pages.groupby('main_category').agg(Websites=('main_category', 'count'),
                                      Local=('Local', 'sum'),
                                      Three_Visits=('EstimatedMonthlyVisits', 'mean'),
                                      Visits=('Engagments.Visits', 'mean'),
                                      Time_Visits=('Engagments.TimeOnSite', 'mean'),
                                      Page_Visit=('Engagments.PagePerVisit', 'mean'),
                                      Small=('IsSmall', 'sum'),
                                      Rank=('GlobalRank.Rank', 'mean'),
                                      Bounce=('Engagments.BounceRate', 'mean'),
                                      Search=('TrafficSources.Search', 'mean'),
                                      Direct=('TrafficSources.Direct', 'mean'),
                                      Social=('TrafficSources.Social', 'mean')).round(2).reset_index().sort_values(by='Websites', ascending=False)

df_pages.groupby('main_category').agg(Websites=('main_category', 'count'),
                                      Local=('Local', 'sum'),
                                      Three_Visits=('EstimatedMonthlyVisits', 'median'),
                                      Visits=('Engagments.Visits', 'median'),
                                      Time_Visits=('Engagments.TimeOnSite', 'median'),
                                      Page_Visit=('Engagments.PagePerVisit', 'median'),
                                      Small=('IsSmall', 'sum'),
                                      Rank=('GlobalRank.Rank', 'median'),
                                      Bounce=('Engagments.BounceRate', 'median'),
                                      Search=('TrafficSources.Search', 'median'),
                                      Direct=('TrafficSources.Direct', 'median'),
                                      Social=('TrafficSources.Social', 'median')).round(2).reset_index().sort_values(by='Websites', ascending=False)

df_pages_m.loc[df_pages_m['main_category']!='news_and_media', 'main_category'] = 'specialized_media'

df_pages_m.groupby('main_category').agg(Mentions=('main_category', 'count'),
                                        Local=('Local', 'sum'),
                                        Papers=('Details Page URL', 'nunique'),
                                        Three_Visits=('EstimatedMonthlyVisits', 'mean'),
                                        Visits=('Engagments.Visits', 'mean'),
                                        Time_Visits=('Engagments.TimeOnSite', 'mean'),
                                        Page_Visit=('Engagments.PagePerVisit', 'mean'),
                                        Small=('IsSmall', 'sum'),
                                        Rank=('GlobalRank.Rank', 'mean'),
                                        Bounce=('Engagments.BounceRate', 'mean'),
                                        Search=('TrafficSources.Search', 'mean'),
                                        Direct=('TrafficSources.Direct', 'mean'),
                                        Social=('TrafficSources.Social', 'mean')).round(2).reset_index()

df_mentions = df_pages_m.groupby(['Details Page URL', 'main_category']).agg(Mentions=('main_category', 'count'),
                                                                            Avg_visits=('Engagments.Visits', 'mean'),
                                                                            Avg_time=('Engagments.TimeOnSite', 'mean'),
                                                                            Avg_pages=('Engagments.PagePerVisit', 'mean'),
                                                                            Avg_rank=('GlobalRank.Rank', 'mean'),
                                                                            Avg_bounce=('Engagments.BounceRate', 'mean'),).round(2).reset_index()

df_mentions.to_csv('data/News/publications_metrics_mean_special.tsv', sep='\t', index=False)

df_pages_m.to_csv('data/News/publications_metrics.tsv', encoding='UTF-8', sep='\t', index=False)
