==> 7_wikipedia_analysis.ipynb <==
# Packages

import pandas as pd

# 1. Data import

## 1.1. Mentions

df = pd.read_csv('data/Wikipedia/mentions.tsv', sep='\t', encoding='UTF-8', dtype={'pageid':'str'})
df.shape

## 1.2. Pages

df_pages = pd.read_csv('data/Wikipedia/pages.tsv', sep='\t', encoding='UTF-8', dtype={'Wikipedia_page_id':'str'})
df_pages.shape

df_pages[df_pages['Wikipedia_page_id'].isin(df.pageid)].shape

df[df['pageid'].isin(df_pages.Wikipedia_page_id)].shape

# 2. Wikipedia pages

df_pages.loc[df_pages['major_topic'].isna(), 'major_topic'] = 'Other'

df_pages.groupby(['major_topic']).agg(Pages=('major_topic', 'count'),
                                      Words=('words', 'mean'),
                                      Languages=('Wikipedia_lg', 'nunique'),
                                      Translations=('langs', 'mean'),
                                      Pageviews=('pageviews', 'mean'),
                                      Editions=('revisions', 'mean'),
                                      References=('references', 'mean')
                                     ).round(2).reset_index()

df_pages.groupby(lambda x: True).agg(Pages=('major_topic', 'count'),
                                     Words=('words', 'mean'),
                                     Languages=('Wikipedia_lg', 'nunique'),
                                     Translations=('langs', 'mean'),
                                     Pageviews=('pageviews', 'mean'),
                                     Editions=('revisions', 'mean'),
                                     References=('references', 'mean')
                                     ).round(2).reset_index()

df_pages.groupby(['major_topic']).agg(Pages=('major_topic', 'count'),
                                      Words=('words', 'median'),
                                      Languages=('Wikipedia_lg', 'nunique'),
                                      Translations=('langs', 'median'),
                                      Pageviews=('pageviews', 'median'),
                                      Editions=('revisions', 'median'),
                                      References=('references', 'median')
                                     ).round(2).reset_index()

df_pages.groupby(lambda x: True).agg(Pages=('major_topic', 'count'),
                                     Words=('words', 'median'),
                                     Languages=('Wikipedia_lg', 'nunique'),
                                     Translations=('langs', 'median'),
                                     Pageviews=('pageviews', 'median'),
                                     Editions=('revisions', 'median'),
                                     References=('references', 'median')
                                     ).round(2).reset_index()

# 3. Wikipedia mentions

df_pages_m = df_pages.merge(df[['pageid', 'user', 'Details Page URL', 'DOI']], how='inner', left_on='Wikipedia_page_id', right_on='pageid').drop('pageid', axis=1)

df_pages_m['local'] = 0
df_pages_m.loc[df_pages_m['Wikipedia_lg'].isin(['es','ca']),'local'] = 1

df_pages_m.groupby('major_topic').agg(Unique_user=('user', 'nunique'),
                                      Papers=('Details Page URL', 'nunique'),
                                      Local=('local', 'sum'),
                                      Mentions=('major_topic', 'count'),
                                      Pageviews=('pageviews', 'mean'),
                                      Edits=('revisions', 'mean')).round(2).reset_index()

df_pages_m.groupby(lambda x: True).agg(Unique_user=('user', 'nunique'),
                                       Papers=('Details Page URL', 'nunique'),
                                       Local=('local', 'sum'),
                                       Mentions=('major_topic', 'count'),
                                       Pageviews=('pageviews', 'mean'),
                                       Edits=('revisions', 'mean')).round(2).reset_index()

df_pages_m.groupby('major_topic').agg(Unique_user=('user', 'nunique'),
                                      Local=('local', 'sum'),
                                      Papers=('Details Page URL', 'nunique'),
                                      Mentions=('major_topic', 'count'),
                                      Pageviews=('pageviews', 'median'),
                                      Edits=('revisions', 'median')).round(2).reset_index()

df_pages_m.groupby(lambda x: True).agg(Unique_user=('user', 'nunique'),
                                       Local=('local', 'sum'),
                                       Papers=('Details Page URL', 'nunique'),
                                       Mentions=('major_topic', 'count'),
                                       Pageviews=('pageviews', 'median'),
                                       Edits=('revisions', 'median')).round(2).reset_index()

df_mentions = df_pages_m.groupby(['Details Page URL', 'major_topic']).agg(Mentions=('major_topic', 'count'),
                                                                          Avg_views=('pageviews', 'mean'),
                                                                          Avg_edits=('revisions', 'mean'),
                                                                          Avg_words=('words', 'mean'),
                                                                          Avg_trans=('langs', 'mean')).round(2).reset_index()

df_mentions.to_csv('data/Wikipedia/publications_metrics_mean.tsv', sep='\t', index=False)

df_mentions = df_pages_m.groupby(['Details Page URL']).agg(Mentions=('Details Page URL', 'count'),
                                                           Avg_views=('pageviews', 'mean'),
                                                           Avg_edits=('revisions', 'mean'),
                                                           Avg_words=('words', 'mean'),
                                                           Avg_trans=('langs', 'mean')).round(2).reset_index()

df_mentions.to_csv('data/Wikipedia/publications_metrics_mean_all.tsv', sep='\t', index=False)

df_pages_m.to_csv('data/Wikipedia/publications_metrics.tsv', sep='\t', index=False)
